requirements.txt
streamlit
pandas
numpy
plotly
requests
python-dotenv
yfinance
scikit-learn
matplotlib
seaborn
djangorestframework
django-cors-headers
